pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-bootstrap5>=25.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.4",
    "whitenoise>=6.9.0",
]
